Do not pass claimed events further down
This is an issue, for example, when using the SplitScreenPass as moving
the divider between the two splits results in the camera moving along.

`SplitScreenPass::onMouseEvent()` returns true when the divider is
grabbed, as expected, but Mogwai's `Renderer::onMouseEvent()` does not
check the return value of the `RenderGraph::onMouseEvent()` method and
so will still call the `onMouseEvent()` method on the scene (therefore
resulting in the camera moving along despite the SplitScreenPass pass
claiming the events).

The same issue applies to keyboard events.